refactor: extract the same logic into a function
